split logic into multiple files, added share function
