+++ platform-window.html
@@ -9,6 +9,10 @@
         <link rel="stylesheet" href="styles/frame-styles.css"></link>
         <link rel="stylesheet" href="styles/light-theme.css"></link>
         <script type="module" src="js/platform-window.js"></script>
+        <script type="module" src="js/left-menu.js"></script>
+        <script type="module" src="js/title-bar.js"></script>
+        <script type="module" src="js/layout-form.js"></script>
+        <script type="module" src="js/snapshot-form.js"></script>
     </head>
     <body>
         <div id="of-frame-main">
@@ -17,8 +21,8 @@
                 <left-menu class="hidden"></left-menu>
                 <div class="two-sided-container">
                     <div id="layout-container" class="face"></div>
-                    <layout-menu class="face hidden"></layout-menu>
-                    <snapshot-menu class="face hidden"></snapshot-menu>
+                    <layout-form class="face hidden"></layout-form>
+                    <snapshot-form class="face hidden"></snapshot-form>
                 </div>
             </div>
         </div>
